unify mqtt configuration to common
